RUMM-779 Reorganize test scenarios so it is easier to add new ones
